refactor(backend): rename task graph package to task context
