+++ src/backend/specs/2026-08-09-task-goal-driven-task-runner-bbs/spec.md
@@ -337,7 +337,7 @@ bot1 `claim` 长活 → `attach` 节点 → 周期性 `bbs/result`(`output_patch
 
 ### 13.1 现有 task API(`adapters/http/task/router.py`,prefix `/openapi/v1/task`)
 - 读面:`POST /openapi/v1/collaboration/tasks/execute`、`GET /openapi/v1/collaboration/tasks/dashboard`(含 `?task_id=`、`?node_id=` 子树投影)、`GET /openapi/v1/collaboration/tasks/list`(含 `?status=`)、`POST /api/v1/collaboration/tasks/callback/report`。另有 `task_callback_router`(prefix `/api/v1/collaboration/tasks/callback`,`/workflow_start`/`/workflow_result`/`/node_start`/`/node_result`,HMAC 鉴权)——BBS 不走此(其为 workflow 引擎回调)。
-- 权威:`domain/models.py`(6 态 `Status`、`AcceptanceVerdict`、`TaskNode`、`TaskExecutionGraph`、`RuntimeInfo`、`TaskNodePatch`、`TaskGraphPatch`、`PlanResult`、`TaskCallbackData`);`task_graph/task_graph_service.py`(SSOT 写口 `initialize_graph`/`add_task_nodes`/`update_task_node_info`/`update_task_graph_info` + 转移表 `_ACCEPTANCE_TRANSITIONS`/`_DIRECT_TRANSITIONS`,`FAILED→{PENDING,HUNG}`、`HUNG` 终态);`task_center/engine.py`(`ExecutionEngine.on_execute/on_start/on_report/on_miss/on_harness` + 本 spec 新增 `on_bbs_report` + `_hung_and_escalate`/`_bump_loop_round`/HUNG 冒泡);`task_harness/harness.py`(`TaskHarness` SLA backstop);`task_center/task_service.py`(`TaskService` facade:`execute`/`get_task_dashboard`/`list_tasks`/`.callback`)。
+- 权威:`domain/models.py`(6 态 `Status`、`AcceptanceVerdict`、`TaskNode`、`TaskExecutionGraph`、`RuntimeInfo`、`TaskNodePatch`、`TaskGraphPatch`、`PlanResult`、`TaskCallbackData`);`task_context/task_graph_service.py`(SSOT 写口 `initialize_graph`/`add_task_nodes`/`update_task_node_info`/`update_task_graph_info` + 转移表 `_ACCEPTANCE_TRANSITIONS`/`_DIRECT_TRANSITIONS`,`FAILED→{PENDING,HUNG}`、`HUNG` 终态);`task_center/engine.py`(`ExecutionEngine.on_execute/on_start/on_report/on_miss/on_harness` + 本 spec 新增 `on_bbs_report` + `_hung_and_escalate`/`_bump_loop_round`/HUNG 冒泡);`task_harness/harness.py`(`TaskHarness` SLA backstop);`task_center/task_service.py`(`TaskService` facade:`execute`/`get_task_dashboard`/`list_tasks`/`.callback`)。
 
 ### 13.2 本 spec 加法(契约面,HOW 留 plan)
 - 新路由:`POST /api/v1/collaboration/tasks/bbs/claim`、`POST /api/v1/collaboration/tasks/bbs/attach`、`POST /api/v1/collaboration/tasks/bbs/result`。
