+++ src/backend/specs/2026-08-09-task-goal-driven-task-runner/plan.md
@@ -425,7 +425,7 @@ def test_dispatch_returns_one_bool_per_node():
 
 
 def test_runner_falls_back_to_stub_without_backend():
-    from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+    from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
     from agentclaw.community.core.task.task_runner.runner import TaskRunner
     g = TaskGraphService()
     r = TaskRunner(g)  # 无 execution_backend
@@ -2211,7 +2211,7 @@ from agentclaw.community.core.task.domain.models import (
     AcceptanceCriteria, Context, Goal, Metadata, RuntimeInfo, Status, TaskInfo, TaskNode, TaskSpec,
 )
 from agentclaw.community.core.task.task_center.task_service import TaskService
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 from agentclaw.community.core.task.task_runner.integration import build_integration
 from agentclaw.community.core.task.task_runner.integration.double.double_context_provider import _DoubleSink
