+++ src/backend/specs/2026-08-21-task-execute-task-type-branching/plan.md
@@ -548,7 +548,7 @@ from agentclaw.community.core.task.domain.requests import (
 )
 from agentclaw.community.core.task.task_center.engine import CoopGroupStart
 from agentclaw.community.core.task.task_center.task_service import TaskService
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 class _SqliteDB:
