+++ src/backend/tests/community/core/task/e2e/test_e2e.py
@@ -33,7 +33,7 @@ from agentclaw.community.core.task.task_dispatch.strategies import (
     SearchOutcome,
     SearchResult,
 )
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 # ===== domain helpers =====
