+++ src/backend/tests/community/core/task/singlebox_e2e/bbs/test_bbs_relay_e2e.py
@@ -45,7 +45,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskInfo,
     TaskSpec,
 )
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 from agentclaw.community.core.task.task_harness.harness import TaskHarness
 
 pytestmark = pytest.mark.skipif(
