+++ src/backend/tests/community/core/task/task_center/test_engine_bbs_trigger.py
@@ -29,7 +29,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskSpec,
 )
 from agentclaw.community.core.task.task_center.engine import ExecutionEngine
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 # ===== domain helpers (mirrors test_engine.py minimal setup) =====
