+++ src/backend/tests/community/core/task/task_center/test_execute_persist.py
@@ -21,7 +21,7 @@ from agentclaw.community.core.task.domain.requests import (
     RequestTaskSpec, TaskInfoRequest,
 )
 from agentclaw.community.core.task.task_center.task_service import TaskService
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 from agentclaw.community.core.task.repository.types import TaskInfoRecord
