+++ src/backend/tests/community/core/task/task_dispatch/test_dispatcher.py
@@ -27,7 +27,7 @@ from agentclaw.community.core.task.task_dispatch.strategies import (
     SearchOutcome,
     SearchResult,
 )
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 def _run(coro):
