+++ src/backend/tests/community/core/task/task_runner/test_runner.py
@@ -25,7 +25,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskSpec,
 )
 from agentclaw.community.core.task.task_dispatch.protocols import GroupFormation
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 from agentclaw.community.core.task.task_runner.runner import TaskRunner
