+++ src/backend/tests/community/core/task/test_bbs_claim.py
@@ -16,7 +16,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskNode,
     TaskSpec,
 )
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 def _ti(tid="p1"):
