+++ src/backend/tests/community/core/task/test_bbs_claim_route.py
@@ -25,7 +25,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskSpec,
 )
 from agentclaw.community.core.task.domain.errors import TaskStateError
-from agentclaw.community.core.task.task_graph.task_graph_service import TaskGraphService
+from agentclaw.community.core.task.task_context.task_graph_service import TaskGraphService
 
 
 class _StubDiscoverModule(Module):
