+++ src/backend/tests/community/endpoints/test_task_endpoints.py
@@ -32,7 +32,7 @@ from agentclaw.community.core.task.domain.models import (
     TaskNodePatch,
     TaskSpec,
 )
-from agentclaw.community.core.task.task_graph.task_graph_service import (
+from agentclaw.community.core.task.task_context.task_graph_service import (
     TaskGraphService,
 )
 from tests.community.framework import (
